Squashed commit of the following:
commit 6891e5c8a85e489570cacee037b5f6db43fd2932
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 30 15:32:49 2020 +0200

    maed it more silent

commit b0c9b966451e4157dc5c80560b5948a7e3d7d0b6
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 30 15:23:42 2020 +0200

    made rsync check checksums

commit 5f9eea6f0afb32a605b791dbdd14eb98a9ab0c83
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 30 15:12:58 2020 +0200

    made it auto sense kitlot

commit bf5ee980d4101b2953239df71a00ca3cf74afc9d
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 2 14:11:06 2020 +0200

    added renv sources

commit f5c7f8facfdbf664663a463ded804d7f20cc9528
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 2 13:56:39 2020 +0200

    removed renv from lock file

commit 0260e2cebe836c6054e7e59d03cd08042d3494de
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 2 11:16:55 2020 +0200

    ignore

commit 08f20125a0a9c8317676c00fabedb24604f7c022
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 2 11:13:44 2020 +0200

    removed packrat package

commit cd56a294240220bab2056a0be3657a8fbe915009
Author: Torbjørn Lindahl <torbjorn.lindahl@gmail.com>
Date:   Tue Jun 2 11:10:28 2020 +0200

    moved to renv

+++ watch.sh
@@ -16,7 +16,7 @@ inotifywait -r -m $WATCHED_DIR --format '%e %f' ~/git/R-packages/ga.software.dd 
 
   echo $event $file
 
-  if [ $file != '.#'* ]; then
+  if [[ $file != '.#'* && $file != 'index.lock' ]]; then
       kill $PID
       start
   fi
